Initial support for collection 2 SR (and collection 1 TOA)
I added separate classes for each of the Landsat collections as well as a generic Landsat class that can be passed an image ID.  I started an Model class that will eventually hold all of the LAI calculation functions, but is currently just a wrapper to the getLAIImage function.

I am currently scaling the collection 2 SR and colleciton 1 TOA images by 10000 to make them match the collection 1 SR images, but I think in the long term we should adjust the code to assume that the SR images were all scaled to 0-1.

Eventually I would also like to move Model class and code to code to a separate file but it was easier to get the inheritance working without worrying about imports.

I'm not sure why the getLAIforBiome tests are all failing.  I don't think I changed any of that code but I need to switch back to Master and double check.