Miscellaneous changes and updates
Dropped the camel case on the function names and other

Updated example notebooks

Switched to the cloud assets path for the training dataset.

Renamed the prepped QA band to just "qa" to make is more distinct from "pixel_qa" and "QA_PIXEL".

Switched to using the utils get_info call for all of the test .getInfo() calls.